refactor: ProjectDetailView を分離（コロケーション）
- app/projects/[id]/page.tsx → ProjectDetailView.tsx に移動
- 新しい page.tsx は View を呼び出すだけ（5行）
- useParams はそのまま View 内で使用
